sitemap_pro/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
playwright>=1.40.0
httpx>=0.26.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
Pillow>=10.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
python-dotenv>=1.0.0
networkx>=3.0
# Text Extraction & Comparison (Legacy/Optional)
PyMuPDF>=1.23.0
diff-match-patch>=20230430
python-multipart>=0.0.6
